// frontend/src/map/common/types.ts
export interface ActiveWaypoint {
  title: string;
  description?: string;
  selected?: boolean;
  addresslnglat?: [number, number];
  sourcelnglat?: [number, number];
  displaylnglat: [number, number];
  key: number;
  addressindex: number;
}

export type ActiveWaypoints = ActiveWaypoint[];

export type Settings = Record<
  'costing' | 'directions',
  {
    [key in keyof PossibleSettings]: PossibleSettings[key];
  }
>;

export type BicycleType = 'Hybrid' | 'Road' | 'City' | 'Cross' | 'Mountain';

export interface PossibleSettings {
  maneuver_penalty: number;
  country_crossing_penalty: number;
  country_crossing_cost: number;
  length: number;
  width: number;
  height: number;
  weight: number;
  axle_load: number;
  hazmat: boolean;
  use_highways: number;
  use_tolls: number;
  use_ferry: number;
  ferry_cost: number;
  use_living_streets: number;
  use_tracks: number;
  private_access_penalty: number;
  ignore_closures: boolean;
  ignore_restrictions: boolean;
  ignore_access: boolean;
  closure_factor: number;
  service_penalty: number;
  service_factor: number;
  exclude_unpaved: boolean;
  shortest: boolean;
  exclude_cash_only_tolls: boolean;
  bicycle_type: BicycleType;
  cycling_speed: number;
  use_roads: number;
  use_hills: number;
  avoid_bad_surfaces: number;
  top_speed: number;
  use_primary: number;
  walking_speed: number;
  walkway_factor: number;
  sidewalk_factor: number;
  alley_factor: number;
  driveway_factor: number;
  step_penalty: number;
  max_hiking_difficulty: number;
  exclude_polygons: GeoJSON.GeoJSON[];
  use_geocoding: boolean;
  use_lit: number;
  axle_count: number;
  fixed_speed: number;
  toll_booth_penalty: number;
  toll_booth_cost: number;
  gate_penalty: number;
  gate_cost: number;
  include_hov2: boolean;
  include_hov3: boolean;
  include_hot: boolean;
  transit_start_end_max_distance: number;
  transit_transfer_max_distance: number;
  disable_hierarchy_pruning: boolean;
  use_trails: number;
  denoise: number;
  generalize: number;
  alternates: number;
  [key: string]: string | number | boolean | GeoJSON.GeoJSON[] | undefined;
}